Integrate SpermModel5GPU to the sample projects
After updating the original SpermModel by Mark Burkitt to execute
under FlameGPU 1.0.4, it makes sense to add the MSVS2013 project to the
examples folder.

Changes are also introduced to the FlameGPU schemas. These extend the
framework with the custom functionality developed by Burkitt.
Extended documentation should be available in the project wiki.